Removed unmaintained code from the repo. Updated wrappers so that if they're turned off then that mainly affects auto-defined Granary wrappers, but not client-defined wrappers. Turned on wrappers, even for full-kernel. Added WRAP_FUNCTION_ONCE, for use by clients using wrappers to interpose on kernel functions, but also wanting to instrument those functions. This is less flaky than the patch wrapper approach. Fixed some watchpoint issues having to do with operand sizes and wrappers. Made the default wrapping approach for RUNNING_IN_APP during whole-kernel be to not wrap, i.e. apply as a zero enumeration constant. Re-enabled page protecting of module code.

+++ clients/watchpoints/instrument.h
@@ -183,7 +183,7 @@ namespace client {
 
 
             /// The number of operands that need to be instrumented.
-            unsigned num_ops;
+            uint8_t num_ops;
 
 
             /// Track the carry flag.
